+++ src/components/Header/styles.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Container = styled.header`
+  width: 100%;
+  height: 250px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  nav {
+    margin: 50px 0;
+    a img {
+      width: 300px;
+    }
+  }
+`;