+++ src/pages/_layouts/default/styles.js
@@ -2,6 +2,6 @@ import styled from 'styled-components';
 import background from '../../../assets/background.png';
 
 export const Wrapper = styled.div`
-  height: 100%;
   background: url(${background}), black;
+  margin: 0 auto;
 `;